python-graphql-api/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
graphene==3.3
graphql-core==3.2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==3.0.1
PyJWT==2.8.0
boto3==1.34.34
